+++ src/main/java/com/yisylvie/createtoolboxtooltip/access/PreviewCategoryAccess.java
@@ -1,0 +1,7 @@
+package com.yisylvie.createtoolboxtooltip.access;
+
+// The duck interface being implemented onto PreviewCategory
+// Too old to be a duckling! Quack, quack!
+public interface PreviewCategoryAccess {
+	boolean createtoolboxtooltip$getEnableOnlyToolboxes();
+}