+++ src/main/java/com/yisylvie/createtoolboxtooltip/createToolboxTooltip.java
@@ -2,13 +2,13 @@ package com.yisylvie.createtoolboxtooltip;
 
 import com.misterpemodder.shulkerboxtooltip.api.neoforge.ShulkerBoxTooltipPlugin;
 
-import net.minecraft.resources.ResourceLocation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.Mod;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import net.minecraft.resources.ResourceLocation;
 
 @Mod(createToolboxTooltip.ID)
 public final class createToolboxTooltip {
@@ -16,9 +16,9 @@ public final class createToolboxTooltip {
     public static final String NAME = "Create Toolbox Tooltip";
     public static final Logger LOGGER = LoggerFactory.getLogger(NAME);
 
-    public static ResourceLocation id(String path) {
-		return ResourceLocation.fromNamespaceAndPath(ID, path);
-    }
+    // public static ResourceLocation id(String path) {
+	// 	return ResourceLocation.fromNamespaceAndPath(ID, path);
+    // }
 
 	public createToolboxTooltip() {
 		// plugin registration